// dist/node/runtime/node/community/page-generator.d.ts
export type PageOptions = {
    excludeRemovedComments: boolean;
    excludeDeletedComments: boolean;
    excludeCommentPendingApproval: boolean;
    excludeCommentWithApprovedFalse: boolean;
    excludeCommentsWithDifferentCommunityAddress: boolean;
    commentUpdateFieldsToExclude?: (keyof CommentUpdateType)[];
    parentCid: string | null;
    preloadedPage: PostSortName | ReplySortName;
    baseTimestamp: number;
    firstPageSizeBytes: number;
};